\name{FRB-internal}
\alias{Seinfs_multireg}
\alias{MMeinfs_multireg}
\alias{GSeinfs_multireg}
\alias{Seinfs_pca}
\alias{MMeinfs_pca}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Internal functions for the package FRB}
\description{
These functions are not intended to be used by the user.
}
%\usage{
%Seinfs_multireg(X,Y,ests=Sest_multireg(X, Y))
%MMeinfs_multireg(X,Y,ests=MMest_multireg(X, Y))
%GSeinfs_multireg(X,Y,ests=GSest_multireg(X, Y))
%
%Seinfs_pca(Y, ests = Sest_loccov(Y))
%MMeinfs_pca(Y, ests = MMest_loccov(Y))
%}
%- maybe also 'usage' for other objects documented here.
