% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frb.R
\name{frb}
\alias{frb}
\title{Fast and Robust Bootstrap}
\usage{
frb(lmrob.object, nboot = 1000, return.coef = FALSE, return.indices = FALSE)
}
\arguments{
\item{lmrob.object}{a robust linear regression fit as returned by \code{lmrob} of \code{package:robustbase}}

\item{nboot}{an integer indicating the number of bootstrap samples to use}

\item{return.coef}{a logical (boolean) value indicating whether to return a matrix containing
the \code{nboot} vectors of bootstrapped regression coefficient estimates.}

\item{return.indices}{a logical (boolean) value indicating whether to return a matrix containing
the vectors of indices corresponding to each bootstrap sample used.}
}
\value{
A list with the following possible components:
\item{coef}{If the argument \code{return.coef == TRUE}, a matrix with 
\code{nboot} rows containing the bootstrapped robust regression estimates}
\item{var}{The sample covariance matrix of the bootstrapped regression estimates. This
is an estimate for the asymptotic covariance matrix of the robust regression estimator.}
\item{indices}{If the argument \code{return.indices == TRUE}, a matrix with 
\code{nboot} rows containing the indices of the bootstrap samples used.}
}
\description{
This function computes the Fast and Robust Bootstrap for linear models
using etc. etc. It relies on the function \code{lmrob} of 
\code{package:robustbase}
}
\details{
This function computes the Fast and Robust Bootstrap for linear models
using etc. etc. It relies on the function \code{lmrob} of 
\code{package:robustbase}
}
\examples{
library(robustbase)
a <- lmrob(LNOx ~ LNOxEm + sqrtWS, data=NOxEmissions)
set.seed(123)
tmp <- frb(lmrob.object=a, nboot=1000, return.coef=FALSE)
sqrt(diag(tmp$var))
# compare with SE's based on the asymptotic approximation
sqrt(diag(summary(a)$cov))
}
\references{
Salibian-Barrera M, Zamar RH, Bootstrapping robust estimates of regression,
Annals of Statistics, (2002) 30:556-582. \doi{10.1214/aos/1021379865}
Salibian-Barrera M, Van Aelst S, Willerns G., Fast and robust bootstrap, 
Statistical Methods and Applications, (2008) 17:41-71. \doi{10.1007/s10260-007-0048-6}
}
\seealso{
\code{package:robustbase}
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
